+++ internal/model/order.go
@@ -6,7 +6,7 @@ type Order struct {
 	ID         *int      `json:"-"`
 	User       *User     `json:"-"`
 	Number     string    `json:"number"`
-	Accrual    *int      `json:"accrual,omitempty"`
+	Accrual    *float32  `json:"accrual,omitempty"`
 	Status     string    `json:"status"`
 	UploadTime time.Time `json:"uploaded_at,omitempty"`
 }
