+++ internal/model/withdraw.go
@@ -4,7 +4,7 @@ import "time"
 
 type Withdraw struct {
 	Order       string    `json:"order"`
-	Sum         int       `json:"sum"`
+	Sum         float32   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at,omitempty"`
 	User        User      `json:"-"`
 }
